fix: address review feedback for typed Config module
- main.cpp: fix uint32_t->int narrowing (use uint32_t directly)
- config.hpp: add [[nodiscard]] to Config::load()
- config.cpp: merge parse_uint16/32 into template parse_int<T>
- config.cpp: switch from static functions to anonymous namespace
- config.cpp: start CLI loop from i=1 (skip argv[0] program name)
- config.cpp: output stderr warning on parse failure for known keys
- config_test.cpp: prepend dummy argv[0] for CLI tests

+++ src/main.cpp
@@ -37,7 +37,7 @@ int main(int argc, char** argv) {
 
         int client_fd = *client;
 
-        int max_msg = static_cast<int>(config.max_message_bytes);
+        uint32_t max_msg = config.max_message_bytes;
         std::string log_root_copy = config.log_root;
 
         std::thread([client_fd, metadata, log_root_copy, max_msg] {
